Split hook.py into focused submodules
Create git.py (git plumbing), driver.py (build-driver discovery),
execution.py (build execution/output capture), state.py (shared
data structures/orchestration). Slim hook.py to CLI entry point
with re-exports for backward compatibility.

82/107 tests pass; remaining 25 failures are test mock-patching
across submodule boundaries (see TODO).
